refactor(ide): move symbol ids into semantic facts
