+++ crates/ide/src/definitions.rs
@@ -1,41 +1,20 @@
 use hir::{
-    base_db::intern::Lookup,
-    container::{ContainerId, InContainer, InFile, InModule, InSubroutine},
+    container::{ContainerId, InContainer, InModule},
     db::HirDb,
     file::HirFileId,
-    hir_def::{
-        block::{BlockId, BlockLoc},
-        expr::declarator::DeclId,
-        file::{config::ConfigDeclId, library::LibraryDeclId, udp::UdpDeclId},
-        module::{
-            ModuleId,
-            generate::{GenerateBlockId, GenerateBlockLoc},
-            instantiation::InstanceId,
-            port::NonAnsiPortId,
-        },
-        stmt::StmtId,
-        subroutine::{SubroutineId, SubroutinePortId},
-        typedef::TypedefId,
-    },
     semantics::{Semantics, pathres::PathResolution},
-    source_map::{IsNamedSrc, IsSrc, ToAstNode},
 };
 use smallvec::{SmallVec, smallvec};
-use smol_str::SmolStr;
 use syntax::{
     SyntaxAncestors, SyntaxToken, SyntaxTokenWithParent,
     ast::{self, AstNode},
     has_name::HasName,
-    has_text_range::{HasTextRange, HasTextRangeIn},
     match_ast,
     token::TokenKindExt,
 };
-use utils::{
-    get::{Get, GetRef},
-    impl_from,
-    line_index::TextRange,
-};
+use utils::impl_from;
 
+pub use crate::facts::symbol::DefinitionOrigin;
 use crate::{
     db::root_db::RootDb,
     module_resolution::{
@@ -43,256 +22,6 @@ use crate::{
         resolve_named_port_connection,
     },
 };
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
-pub enum DefinitionOrigin {
-    ModuleId(ModuleId),
-    Config(InFile<ConfigDeclId>),
-    Library(InFile<LibraryDeclId>),
-    Udp(InFile<UdpDeclId>),
-    BlockId(BlockId),
-    GenerateBlockId(GenerateBlockId),
-    SubroutineId(SubroutineId),
-    SubroutinePort(InSubroutine<SubroutinePortId>),
-
-    NonAnsiPort(InModule<NonAnsiPortId>),
-    Decl(InContainer<DeclId>),
-    Typedef(InContainer<TypedefId>),
-    Instance(InModule<InstanceId>),
-    Stmt(InContainer<StmtId>),
-}
-
-impl_from! { DefinitionOrigin =>
-    ModuleId,
-    Config(InFile<ConfigDeclId>),
-    Library(InFile<LibraryDeclId>),
-    Udp(InFile<UdpDeclId>),
-    BlockId,
-    GenerateBlockId,
-    SubroutineId,
-    SubroutinePort(InSubroutine<SubroutinePortId>),
-    NonAnsiPort(InModule<NonAnsiPortId>),
-    Decl(InContainer<DeclId>),
-    Typedef(InContainer<TypedefId>),
-    Instance(InModule<InstanceId>),
-    Stmt(InContainer<StmtId>),
-}
-
-impl DefinitionOrigin {
-    #[inline]
-    pub fn container_id(&self, db: &dyn HirDb) -> ContainerId {
-        match *self {
-            DefinitionOrigin::ModuleId(InFile { file_id, .. }) => file_id.into(),
-            DefinitionOrigin::Config(InFile { file_id, .. }) => file_id.into(),
-            DefinitionOrigin::Library(InFile { file_id, .. }) => file_id.into(),
-            DefinitionOrigin::Udp(InFile { file_id, .. }) => file_id.into(),
-            DefinitionOrigin::BlockId(block_id) => block_id.lookup(db).cont_id,
-            DefinitionOrigin::GenerateBlockId(generate_block_id) => {
-                generate_block_id.lookup(db).cont_id
-            }
-            DefinitionOrigin::SubroutineId(subroutine_id) => {
-                subroutine_id.lookup(db).cont_id.into()
-            }
-            DefinitionOrigin::SubroutinePort(InSubroutine { subroutine, .. }) => {
-                ContainerId::SubroutineId(subroutine)
-            }
-            DefinitionOrigin::NonAnsiPort(InModule { module_id, .. }) => module_id.into(),
-            DefinitionOrigin::Decl(InContainer { cont_id, .. }) => cont_id,
-            DefinitionOrigin::Typedef(InContainer { cont_id, .. }) => cont_id,
-            DefinitionOrigin::Instance(InModule { module_id, .. }) => module_id.into(),
-            DefinitionOrigin::Stmt(InContainer { cont_id, .. }) => cont_id,
-        }
-    }
-
-    pub fn name(&self, db: &dyn HirDb) -> Option<SmolStr> {
-        match *self {
-            DefinitionOrigin::ModuleId(InFile { value, file_id }) => {
-                file_id.to_container(db).get(value).name.clone()
-            }
-            DefinitionOrigin::Config(InFile { value, file_id }) => {
-                file_id.to_container(db).get(value).name.clone()
-            }
-            DefinitionOrigin::Library(InFile { value, file_id }) => {
-                file_id.to_container(db).get(value).name.clone()
-            }
-            DefinitionOrigin::Udp(InFile { value, file_id }) => {
-                file_id.to_container(db).get(value).name.clone()
-            }
-            DefinitionOrigin::BlockId(block_id) => {
-                let BlockLoc { cont_id, src: InFile { value, file_id: _ } } = block_id.lookup(db);
-                let cont = cont_id.to_container(db);
-                value.hir(&cont, &cont_id.to_container_src_map(db))?.name.clone()
-            }
-            DefinitionOrigin::GenerateBlockId(generate_block_id) => {
-                db.generate_block(generate_block_id).name.clone()
-            }
-            DefinitionOrigin::SubroutineId(subroutine_id) => {
-                db.subroutine(subroutine_id).name.clone()
-            }
-            DefinitionOrigin::SubroutinePort(InSubroutine { subroutine, value }) => {
-                db.subroutine(subroutine).ports.get(value.0 as usize)?.name.clone()
-            }
-            DefinitionOrigin::NonAnsiPort(InModule { value, module_id }) => {
-                module_id.to_container(db).get(value).label.clone()
-            }
-            DefinitionOrigin::Decl(InContainer { value, cont_id }) => {
-                cont_id.to_container(db).get(value).name.clone()
-            }
-            DefinitionOrigin::Typedef(InContainer { value, cont_id }) => {
-                cont_id.to_container(db).get(value).name.clone()
-            }
-            DefinitionOrigin::Instance(InModule { value, module_id }) => {
-                module_id.to_container(db).get(value).name.clone()
-            }
-            DefinitionOrigin::Stmt(InContainer { value, cont_id }) => {
-                cont_id.to_container(db).get(value).label.clone()
-            }
-        }
-    }
-
-    pub fn name_range(&self, db: &dyn HirDb) -> Option<InFile<TextRange>> {
-        match *self {
-            DefinitionOrigin::ModuleId(InFile { value, file_id }) => {
-                let range = file_id.to_container_src_map(db).get(value)?.name_range()?;
-                Some(InFile::new(file_id, range))
-            }
-            DefinitionOrigin::Config(InFile { value, file_id }) => {
-                let range = file_id.to_container_src_map(db).get(value)?.name_range()?;
-                Some(InFile::new(file_id, range))
-            }
-            DefinitionOrigin::Library(InFile { value, file_id }) => {
-                let range = file_id.to_container_src_map(db).get(value)?.name_range()?;
-                Some(InFile::new(file_id, range))
-            }
-            DefinitionOrigin::Udp(InFile { value, file_id }) => {
-                let range = file_id.to_container_src_map(db).get(value)?.name_range()?;
-                Some(InFile::new(file_id, range))
-            }
-            DefinitionOrigin::BlockId(block_id) => {
-                let BlockLoc { src: InFile { value, file_id }, .. } = block_id.lookup(db);
-                let range = value.name_range()?;
-                Some(InFile::new(file_id, range))
-            }
-            DefinitionOrigin::GenerateBlockId(generate_block_id) => {
-                let GenerateBlockLoc { src: InFile { value, file_id }, .. } =
-                    generate_block_id.lookup(db);
-                let range = value.name_range()?;
-                Some(InFile::new(file_id, range))
-            }
-            DefinitionOrigin::SubroutineId(subroutine_id) => {
-                let src = subroutine_id.lookup(db).src;
-                Some(InFile::new(src.file_id, src.value.name_or_full_range()))
-            }
-            DefinitionOrigin::SubroutinePort(InSubroutine { subroutine, value }) => {
-                let src = subroutine.lookup(db).src;
-                let tree = db.parse(src.file_id);
-                let func = src.value.to_node(&tree)?;
-                let ports = func
-                    .prototype()
-                    .port_list()
-                    .map(|ports| ports.ports().children().collect::<Vec<_>>())
-                    .unwrap_or_default();
-                let port = ports
-                    .into_iter()
-                    .nth(value.0 as usize)
-                    .and_then(|port| port.as_function_port())?;
-                let declarator = port.declarator();
-                let range = declarator.name()?.text_range_in(declarator.syntax())?;
-                Some(InFile::new(src.file_id, range))
-            }
-            DefinitionOrigin::NonAnsiPort(InModule { value, module_id }) => {
-                let range = module_id.to_container_src_map(db).get(value)?.name_range()?;
-                Some(InFile::new(module_id.file_id, range))
-            }
-            DefinitionOrigin::Decl(InContainer { value, cont_id }) => {
-                let range = cont_id.to_container_src_map(db).get(value)?.name_range()?;
-                Some(InFile::new(cont_id.file_id(db).into(), range))
-            }
-            DefinitionOrigin::Typedef(InContainer { value, cont_id }) => {
-                let range = cont_id.to_container_src_map(db).get(value)?.name_range()?;
-                Some(InFile::new(cont_id.file_id(db).into(), range))
-            }
-            DefinitionOrigin::Instance(InModule { value, module_id }) => {
-                let range = module_id.to_container_src_map(db).get(value)?.name_range()?;
-                Some(InFile::new(module_id.file_id, range))
-            }
-            DefinitionOrigin::Stmt(InContainer { value, cont_id }) => {
-                let range = cont_id.to_container_src_map(db).get(value)?.name_range()?;
-                Some(InFile::new(cont_id.file_id(db).into(), range))
-            }
-        }
-    }
-
-    pub fn range(&self, db: &dyn HirDb) -> Option<InFile<TextRange>> {
-        Some(match *self {
-            DefinitionOrigin::ModuleId(InFile { value, file_id }) => {
-                let range = file_id.to_container_src_map(db).get(value)?.range();
-                InFile::new(file_id, range)
-            }
-            DefinitionOrigin::Config(InFile { value, file_id }) => {
-                let range = file_id.to_container_src_map(db).get(value)?.range();
-                InFile::new(file_id, range)
-            }
-            DefinitionOrigin::Library(InFile { value, file_id }) => {
-                let range = file_id.to_container_src_map(db).get(value)?.range();
-                InFile::new(file_id, range)
-            }
-            DefinitionOrigin::Udp(InFile { value, file_id }) => {
-                let range = file_id.to_container_src_map(db).get(value)?.range();
-                InFile::new(file_id, range)
-            }
-            DefinitionOrigin::BlockId(block_id) => {
-                let BlockLoc { src: InFile { value, file_id }, .. } = block_id.lookup(db);
-                let range = value.range();
-                InFile::new(file_id, range)
-            }
-            DefinitionOrigin::GenerateBlockId(generate_block_id) => {
-                let GenerateBlockLoc { src: InFile { value, file_id }, .. } =
-                    generate_block_id.lookup(db);
-                let range = value.range();
-                InFile::new(file_id, range)
-            }
-            DefinitionOrigin::SubroutineId(subroutine_id) => {
-                let src = subroutine_id.lookup(db).src;
-                let range = src.value.range();
-                InFile::new(src.file_id, range)
-            }
-            DefinitionOrigin::SubroutinePort(InSubroutine { subroutine, value }) => {
-                let src = subroutine.lookup(db).src;
-                let tree = db.parse(src.file_id);
-                let func = src.value.to_node(&tree)?;
-                let ports = func.prototype().port_list()?;
-                let port = ports
-                    .ports()
-                    .children()
-                    .nth(value.0 as usize)
-                    .and_then(|port| port.as_function_port())?;
-                let range = port.syntax().text_range()?;
-                InFile::new(src.file_id, range)
-            }
-            DefinitionOrigin::NonAnsiPort(InModule { value, module_id }) => {
-                let range = module_id.to_container_src_map(db).get(value)?.range();
-                InFile::new(module_id.file_id, range)
-            }
-            DefinitionOrigin::Decl(InContainer { value, cont_id }) => {
-                let range = cont_id.to_container_src_map(db).get(value)?.range();
-                InFile::new(cont_id.file_id(db).into(), range)
-            }
-            DefinitionOrigin::Typedef(InContainer { value, cont_id }) => {
-                let range = cont_id.to_container_src_map(db).get(value)?.range();
-                InFile::new(cont_id.file_id(db).into(), range)
-            }
-            DefinitionOrigin::Instance(InModule { value, module_id }) => {
-                let range = module_id.to_container_src_map(db).get(value)?.range();
-                InFile::new(module_id.file_id, range)
-            }
-            DefinitionOrigin::Stmt(InContainer { value, cont_id }) => {
-                let range = cont_id.to_container_src_map(db).get(value)?.range();
-                InFile::new(cont_id.file_id(db).into(), range)
-            }
-        })
-    }
-}
 
 // Definition may have multiple origins, e.g. non-ansi port
 #[derive(Debug, Clone, PartialEq, Eq)]
